move test to CastTest module
